Switch to using VCR in tests
This means that the tests will run in a predictable way because the
expected responses have been recorded to yaml files and are committed to
the repository. This does mean that the cassettes will need to be
re-recorded every now and then but it avoids test failures on travis
that are just caused by wikipedia responding too slowly or something.