Reduce memory usage of string in hot code path
This line of code gets called a _lot_ when we search recursively through
results for references to other pages.

I've measured the impact of this change with the following code:

    Wikisnakker::Item.find('Q42')

Before this change was made running that line of code resulted in
1,375,689 bytes of memory and 33,206 objects being allocated. Over a
megabyte of memory to retrieve a single record! With this change in
place that code allocates 128,009 bytes of memory and 2,014 objects,
less than 0.2 megabytes!